+++ anyblok_attachment/bloks/attachment/mixin.py
@@ -7,6 +7,7 @@
 # v. 2.0. If a copy of the MPL was not distributed with this file,You can
 # obtain one at http://mozilla.org/MPL/2.0/.
 from anyblok import Declarations
+from anyblok.declarations import hybrid_method
 from anyblok.column import UUID, String
 from anyblok.field import Function
 from .exceptions import NoneValueException, NotLatestException
@@ -33,12 +34,25 @@ class LatestDocument:
             raise NoneValueException("Uuid value is None")
 
         if document.type != 'latest':
-            raise NotLatestException("The document type is not latest")
+            raise NotLatestException(
+                "You try to set a versioned document, this action is "
+                "forbidden"
+            )
 
         self.latest_document_uuid = document.uuid
 
     def del_latest_document(self):
         self.latest_document_uuid = None
+
+    @hybrid_method
+    def is_latest_document(self, document):
+        if document.type != 'latest':
+            raise NotLatestException(
+                "You try to compare the latest document with a versioned "
+                "document, this action is forbidden"
+            )
+
+        return self.latest_document_uuid == document.uuid
 
 
 @Declarations.register(Declarations.Mixin)
@@ -69,3 +83,8 @@ class VersionedDocument:
     def del_versioned_document(self):
         self.versioned_document_uuid = None
         self.versioned_document_version = None
+
+    @hybrid_method
+    def is_versioned_document(self, document):
+        return (self.versioned_document_uuid == document.uuid and
+                self.versioned_document_version == document.version)